+++ src/pages/ProfilesExplorerView/components/SceneExploreServiceFlameGraph/SceneExploreServiceFlameGraph.tsx
@@ -7,9 +7,11 @@ import React from 'react';
 import { FavAction } from '../../domain/actions/FavAction';
 import { SelectAction } from '../../domain/actions/SelectAction';
 import { FiltersVariable } from '../../domain/variables/FiltersVariable/FiltersVariable';
+import { ProfileIdSelectorVariable } from '../../domain/variables/ProfileIdSelectorVariable';
 import { ProfileMetricVariable } from '../../domain/variables/ProfileMetricVariable';
 import { ServiceNameVariable } from '../../domain/variables/ServiceNameVariable/ServiceNameVariable';
 import { GridItemData } from '../SceneByVariableRepeaterGrid/types/GridItemData';
+import { TimeseriesReprocess } from '../SceneLabelValuesTimeseries/domain/events/TimeseriesReprocess';
 import { SceneMainServiceTimeseries } from '../SceneMainServiceTimeseries';
 import { ResolutionBoostExtensionPoint } from './components/ResolutionBoostExtensionPoint';
 import { SceneFlameGraph } from './SceneFlameGraph';
@@ -25,6 +27,7 @@ export class SceneExploreServiceFlameGraph extends SceneObjectBase<SceneExploreS
       key: 'explore-service-flame-graph',
       mainTimeseries: new SceneMainServiceTimeseries({
         item,
+        includeExemplars: true,
         headerActions: (item) => [
           new SceneReactObject({ component: ResolutionBoostExtensionPoint, props: { scene: this } }),
           new SelectAction({ type: 'view-labels', item }),
@@ -54,7 +57,7 @@ export class SceneExploreServiceFlameGraph extends SceneObjectBase<SceneExploreS
   }
 
   initVariables(item: GridItemData) {
-    const { serviceName, profileMetricId, filters } = item.queryRunnerParams;
+    const { serviceName, profileMetricId, filters, profileIdSelector } = item.queryRunnerParams;
 
     if (serviceName) {
       const serviceNameVariable = sceneGraph.findByKeyAndType(this, 'serviceName', ServiceNameVariable);
@@ -66,10 +69,23 @@ export class SceneExploreServiceFlameGraph extends SceneObjectBase<SceneExploreS
       profileMetricVariable.changeValueTo(profileMetricId);
     }
 
+    if (profileIdSelector) {
+      const profileIdSelectorVariable = sceneGraph.findByKeyAndType(
+        this,
+        'profileIdSelector',
+        ProfileIdSelectorVariable
+      );
+      profileIdSelectorVariable.changeValueTo(profileIdSelector);
+    }
+
     if (filters) {
       const filtersVariable = sceneGraph.findByKeyAndType(this, 'filters', FiltersVariable);
       filtersVariable.setState({ filters });
     }
+  }
+
+  reprocessMainTimeseries() {
+    this.state.mainTimeseries?.state.body?.publishEvent(new TimeseriesReprocess({}), true);
   }
 
   // see SceneProfilesExplorer