+++ src/pages/ProfilesExplorerView/components/SceneExploreServiceFlameGraph/domain/events/RemoveProfileIdSelector.ts
@@ -1,0 +1,7 @@
+import { BusEventWithPayload } from '@grafana/data';
+
+export interface RemoveProfileIdSelectorPayload {}
+
+export class RemoveProfileIdSelector extends BusEventWithPayload<RemoveProfileIdSelectorPayload> {
+  public static type = 'remove-profile-id-selector';
+}