+++ src/pages/ProfilesExplorerView/domain/variables/FiltersVariable/FiltersVariable.tsx
@@ -1,34 +1,29 @@
+import { AdHocVariableFilter } from '@grafana/data';
 import { AdHocFiltersVariable, SceneComponentProps, sceneGraph, SceneObject } from '@grafana/scenes';
-import { CompleteFilters, OperatorKind } from '@shared/components/QueryBuilder/domain/types';
+import { CompleteFilters } from '@shared/components/QueryBuilder/domain/types';
 import { QueryBuilder } from '@shared/components/QueryBuilder/QueryBuilder';
+import { buildFilterExpressionParts } from '@shared/components/SavedSearches/utils';
 import { reportInteraction } from '@shared/domain/reportInteraction';
 import { uniq } from 'lodash';
 import React from 'react';
 
 import { useBuildPyroscopeQuery } from '../../useBuildPyroscopeQuery';
 import { ProfilesDataSourceVariable } from '../ProfilesDataSourceVariable';
-import { convertPyroscopeToVariableFilter, isFilterValid } from './filters-ops';
+import { convertPyroscopeToVariableFilter } from './filters-ops';
 
 export class FiltersVariable extends AdHocFiltersVariable {
   static DEFAULT_VALUE = [];
+  private initialFilters?: AdHocVariableFilter[];
 
-  constructor({ key }: { key: string }) {
+  constructor({ key, initialFilters }: { key: string; initialFilters?: AdHocVariableFilter[] }) {
     super({
       key,
       name: key,
       label: 'Filters',
       filters: FiltersVariable.DEFAULT_VALUE,
-      expressionBuilder: (filters) =>
-        filters
-          // after parsing the URL search parameters the filters might end up having an invalid operator, which in turn, will
-          // generate an invalid query that will make the API requests fail - we prevent this to happen by sanitizing the filters here
-          .filter(isFilterValid)
-          .map(({ key, operator, value }) =>
-            operator === OperatorKind['is-empty'] ? `${key}=""` : `${key}${operator}"${value}"`
-          )
-          .join(','),
+      expressionBuilder: (filters) => buildFilterExpressionParts(filters),
     });
-
+    this.initialFilters = initialFilters;
     this.addActivationHandler(this.onActivate.bind(this));
   }
 
@@ -43,6 +38,8 @@ export class FiltersVariable extends AdHocFiltersVariable {
   }
 
   onActivate() {
+    this.setInitialValue();
+
     // VariableDependencyConfig does not work :man_shrug: (never called)
     const dataSourceSub = sceneGraph
       .findByKeyAndType(this, 'dataSource', ProfilesDataSourceVariable)
@@ -53,6 +50,12 @@ export class FiltersVariable extends AdHocFiltersVariable {
     return () => {
       dataSourceSub.unsubscribe();
     };
+  }
+
+  setInitialValue() {
+    if (this.initialFilters && this.initialFilters.length > 0) {
+      this.setState({ filters: this.initialFilters });
+    }
   }
 
   onChangeQuery = (query: string, filters: CompleteFilters) => {